+++ examples/cli_object.py
@@ -9,10 +9,12 @@ class GIT(object):
 
     @hug.object.cli
     def push(self, branch="master"):
+        """Push the latest to origin"""
         return "Pushing {}".format(branch)
 
     @hug.object.cli
     def pull(self, branch="master"):
+        """Pull in the latest from origin"""
         return "Pulling {}".format(branch)