+++ hug/api.py
@@ -469,9 +469,14 @@ class CLIInterfaceAPI(InterfaceAPI):
         self._output_format = formatter
 
     def __str__(self):
-        return "{0}\n\nAvailable Commands:{1}\n".format(
-            self.api.doc or self.api.name, "\n\n\t- " + "\n\t- ".join(self.commands.keys())
-        )
+        output = "{0}\n\nAvailable Commands:\n\n".format(self.api.doc or self.api.name)
+        for command_name, command in self.commands.items():
+            command_string = " - {}{}".format(
+                command_name, ": " + str(command).replace("\n", " ") if str(command) else ""
+            )
+            output += command_string[:77] + "..." if len(command_string) > 80 else command_string
+            output += "\n"
+        return output
 
 
 class ModuleSingleton(type):